+++ src/components/text-field-unit/input/index.ts
@@ -1,4 +1,4 @@
-import { makeStyleSheet } from "../../../styles";
+import { makeStyleSheet } from "../../styles";
 import styles from "./styles.css?inline";
 
 export class SpTextFieldXLargeInput extends HTMLElement {
